Results UI improvements
I spent the first couple days of this week using QL for VS Code in anger trying to track down some IR generation problems on a Linux snapshot. Everything worked pretty well, but there were a few UI improvements that I had to fix to save my sanity:

The results view now has a dropdown to select which result table to display in a multi-table result set.

The results view is now styled based on VS Code theme colors, so the alternating shading on the rows is usable on other color themes. There are still some hard-coded colors for borders and whatnot, but we can worry about those later.

The `EvaluationInfo` object now keeps info about the database and its source archive location, which is then used to resolve locations to file paths. Links in the results view are now clickable.

If the output of a query matches the schema for a "problem" query, the alerts are displayed in the "Problems" panel in VS Code, in addition to the QL results view. This gives squiggly underlines in the text editor, grouping by file, and generally fits in with the way errors and warning work in VS Code. This is similar to how we have a separate alerts view and table view in QL for Eclipse.

The types emitted by the BQRS parsing are now a bit easier to use.