Fix failure to load extension
The extension was failing to load when installed via .vsix, but worked when run under the debugger. The issue was that `ql-vscode` had `fs-extra` as a `devDependency`, but used it at runtime. When launched under the debugger, the extension is run from the original source directory, where `node_modules` includes dev dependencies. When we package the extension into a .vsix, we only include production dependencies.

I fixed the dependency, and then updated `launch.json` so that we now run the extension from the deployment directory instead of the source directory, so we get exactly the same `node_modules` that we would if installed via .vsix. This required a bit of fiddling with source map paths to make debugging still allow source-level breakpoints and stepping.