Handle snapshots that were built on Windows
Source file paths from Linux-built snapshots are absolute paths starting with `/`. Source file paths on Windows start with a drive letter. To make these paths relative to the source archive root, we need to do a bit of normalization to ensure that we have exactly one slash separator, and that the `:` from the drive letter is replaced with `_`.