Improve BQRS parsing performance by 20x
Split the BQRS parsing code out into separate packages so that other extensions can use it:
- `semmle-bqrs`: The actual BQRS parsing code
- `semmle-io`: Platform-independent (Node.js vs. brower) I/O abstractions
- `semmle-io-node`: Node.js-specific implementation of some I/O interfaces

Change the transpiler target for TypeScript to `es2017`, which has more efficient native support for `async`/`await`.

Instead of parsing the BQRS file in the extension and then serializing it to the results view, just pass the file path to the results view and do the parsing there. The results view code parses the file into a minimal JavaScript object model, which is then rendered via the existing React components. The root component has to do a little work to get async updates to its state handled correctly. Results parsing time for a query with ~3800 results has gone done from ~7 seconds to ~350ms.

When showing the results of `@kind problem` queries in the "Problems" pane of VS Code, we now use a new reflection-driven parser that maps the results to annotated TypeScript types. This is overkill for our one use case, but I plan to use this for several other queries in another extension. See `problem-query-results.ts` for an example of the annotations.

The digester is now a single class that consumes a sequence of buffers and allows async reads of individual data values. The only complicated code here is where the read of a single value has to span multiple buffers. In that case, we use a temporary "seam buffer" to concatenate the fragments from multiple buffers so that we can parse the data as a single contiguous buffer.