Activate extension under various different conditions.
This is necessary, for example, if the user opens vscode and tries to
add a database (which would trigger `onCommand:ql.chooseDatabase` if
they do it through the command palette, and `onView:qlDatabases` and
`onCommand:qlDatabases.chooseDatabase` if they do it through the ui)
before selecting a buffer containing ql code (which would trigger
`onLanguage:ql`).

Some of these conditions are probably redundant in the sense that I
believe it to be impossible to trigger them without first triggering
others already in the list, but it seems preferable to get coverage
rather than have a minimal list, since there's not any cost to
ensuring the extension is activated more than once.

Documentation for these activation events can be found at
https://code.visualstudio.com/api/references/activation-events