Switch from using fragment to authority
vscode does not reliably distinguish between workspace folders that
differ only in the fragment (or indeed query) portion of their
defining uri. For example, if you `.updateWorkspaceFolders` to add a
second workspace that resembles an existing one with a different
sourceArchive `.zip` in its fragment, it will silently refuse.
However, authority seems to be respected. On the other hand, we wish
to encode filepaths into the authority, which means slashes, which
must be escaped.