Fix conversion of webview URIs back into filesystem URIs on Windows.
This is used when populating the Problems view, since we receive a BQRS result path from the webview.
The path component of the webview URI is the filesystem path, but it needs to be parsed into a filesystem URI before it can be used.
Attempting to call `fsPath` directly on the webview URI leads to a doubling of the drive letter prefix on Windows.
Factor the conversions into functions, and add an integration test for them.