Adapt source archive URI encoding to work on Windows.
URIs with an authority component must have a path component starting with `/`
(according to https://tools.ietf.org/html/rfc3986#section-3.3).
On Windows, the source archive zip path usually starts with the drive letter,
so add a leading slash if necessary.

To correctly extract the original source archive zip path from the encoded URI,
change the form of the authority component slightly so that it knows both where the
source archive zip path starts and ends in the encoded URI.

The authority is now of the form `startIndex-endIndex` where
- the source archive zip path is `uri.path[startIndex..endIndex)`
- the source file path within the zip is `uri.path[endIndex..uri.path.length)`.

This fixes URI encoding integration tests on Windows, and adds a few more.