Fix handling of RAM settings.
Remove the default value for the setting `codeQL.runningQueries.memory`, allowing it to be empty.
If a value is set, pass it to `codeql resolve ram`, so that CodeQL uses it as the limit.
(The value was accidentally ignored; this is fixed here.)
If no value is set, `codeql resolve ram` will work out the RAM limit based on available memory.