[WIP] Add pagination to queries
This change adds a new section at the bottom of the results table
that instructs a user to load more results.

There is the beginning of adding functionality to call into the
cli to grab the extra results, but I need some direction before
working on this.

Closes #277.

This builds on top of #276. The commits are just cherry-picked, so
there will need to be some reworking once #276 is merged.