Add telemetry for commands
This commit adds telemetry capturing for command execution. The data
captured is only the command id. We also capture errors thrown by any
command execution.

There is a new config setting added that controls whether or not
telemetry should be sent. This setting AND the global setting must
be enabled in order for telemetry to be sent.

Note that the global setting is handled inside the extension by default.