Add open type support from AST Viewer
We make the assumption that AST items whose label start with a bracketed 
name refer to a Code QL AST type.

We would like to use the language server's open type command, but that 
requires a file URI and a text location in that file.

Instead, we use a file search. This is an experimental feature. Let's 
see if it is useful.