Add a timeout to waiting for the user to click on telemetry dialog
If user clicks the 'x', then the dialog goes away without its promise 
ever being resolved. This will prevent the extension from being started.

This commit uses `Promise.race` to set a 10s timeout for the dialog to
disappear and the extension to start anyway. If this happens, the
dialog will appear on next startup and will continue to appear until
the user actively clicks 'Yes' or 'No.