Move some database commands to overflow menu
This commit moves the "Add *" and "Download *" database commands
from the buttons on top and into an overflow menu.

I believe this is easier to read since the icons were never fully
understandable and required hovering to read.

Also, renamed the commands to be more consistent. Now all begin
with "Add Database from".