Draft: Always load databases from extension-controlled locations.
This commit modifies the extension so that when user opens a database from a
directory, the extension first copies the directory to a path that is managed
by the extension itself and then loads its contents. This change is part of the
solution for the problem of CodeQL tests modifying testproj directories behind
the query server's back.